third_party/WebKit/LayoutTests/external/wpt/fetch/http-cache/partial-expected.txt

This is a testharness.js-based test.
FAIL HTTP cache stores partial content and reuses it.assert_less_than: Response used expected a number less than 2 but got 2
FAIL HTTP cache stores complete response and serves smaller ranges from it.assert_equals: Response status expected 200 but got 206
FAIL HTTP cache stores partial response and serves smaller ranges from it.assert_less_than: Response used expected a number less than 2 but got 2
FAIL HTTP cache stores partial content and completes it.assert_equals: expected (string) "bytes=5-" but got (undefined) undefined
Harness: the test ran to completion.
